Refactor Modals to Dialogs in Kanban app
Renamed 'Modal' instances to 'Dialog' to better align with the coding conventions and improved the codes readability. This refactor includes adding new functionality to Dialog components, like the close dialog action when a new tag is created. Also, some code cleaning and reordering has been performed in the UI components.