Update package dependencies across multiple modules
The commit updates package dependencies throughout various modules, including the bumping of "lucide-react" from "^0.402.0" to "^0.403.0", as well as "@types/node" from "^20.14.6" to "^20.14.10". The package "@typescript-eslint/eslint-plugin" and "@typescript-eslint/parser" have also been upgraded from "^7.15.0" to "^7.16.0". This ensures that the latest functionalities and bug fixes from these packages are utilized across the project.